break out a script to run an irl policy
