+++ scripts/run_irl_policy.py
@@ -1,0 +1,37 @@
+from atari_irl import utils, environments, irl
+import pickle
+from arguments import add_atari_args, add_trajectory_args, add_irl_args
+import argparse
+import tensorflow as tf
+from sandbox.rocky.tf.envs.base import TfEnv
+
+
+def run_irl_policy(args):
+    with utils.TfContext():
+        with utils.EnvironmentContext(
+                env_name=args.env,
+                n_envs=args.num_envs,
+                seed=args.seed,
+                **environments.one_hot_atari_modifiers
+        ) as context:
+            envs = environments.VecGymEnv(context.environments)
+            envs = TfEnv(envs)
+
+            policy = irl.make_irl_policy(
+                irl.policy_config(args),
+                envs=envs,
+                sess=tf.get_default_session()
+            )
+            policy.restore_param_values(args.irl_policy_file)
+            policy.show_run_in_gym_env(context.environments)
+
+
+if __name__ == '__main__':
+    parser = argparse.ArgumentParser(
+        formatter_class=argparse.ArgumentDefaultsHelpFormatter
+    )
+    add_atari_args(parser)
+    add_irl_args(parser)
+
+    args = parser.parse_args()
+    run_irl_policy(args)
